Flatten taskmanagement package
