+++ walking-skeleton/src/main/java/com/example/application/taskmanagement/TaskService.java
@@ -1,14 +1,11 @@
-package com.example.application.taskmanagement.service;
+package com.example.application.taskmanagement;
 
-import com.example.application.taskmanagement.domain.Task;
-import com.example.application.taskmanagement.domain.TaskRepository;
 //#if ui.framework == "hilla"
 import com.vaadin.flow.server.auth.AnonymousAllowed;
 import com.vaadin.hilla.BrowserCallable;
 //#endif
 import org.jspecify.annotations.Nullable;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.access.prepost.PreAuthorize;
 //#if ui.framework == "flow"
 import org.springframework.stereotype.Service;
 //#endif
@@ -23,10 +20,8 @@ import java.util.List;
 //#endif
 //#if ui.framework == "hilla"
 @BrowserCallable
-// Until https://github.com/vaadin/hilla/issues/3271 is fixed, @PreAuthorize needs to be combined with @AnonymousAllowed
 @AnonymousAllowed
 //#endif
-@PreAuthorize("isAuthenticated()")
 public class TaskService {
 
     private final TaskRepository taskRepository;
@@ -43,9 +38,7 @@ public class TaskService {
         if ("fail".equals(description)) {
             throw new RuntimeException("This is for testing the error handler");
         }
-        var task = new Task();
-        task.setDescription(description);
-        task.setCreationDate(clock.instant());
+        var task = new Task(description, clock.instant());
         task.setDueDate(dueDate);
         taskRepository.saveAndFlush(task);
     }
