+++ walking-skeleton/src/main/java/com/example/application/taskmanagement/package-info.java
@@ -11,5 +11,14 @@
  * directly to the application package.
  * </p>
  */
+@NullMarked
+//#if ui.framework == "hilla"
+@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
+//#endif
 package com.example.application.taskmanagement;
 // TODO Remove this package once you have added real features
+
+import org.jspecify.annotations.NullMarked;
+//#if ui.framework == "hilla"
+import org.springframework.lang.NonNullApi;
+//#endif
