--- walking-skeleton/src/main/java/com/example/application/taskmanagement/service/package-info.java
@@ -1,17 +1,0 @@
-/**
- * This package contains the application services of the Task Management sample feature.
- * <p>
- * You can add as many application services as you want to this package, as long as they belong to the same feature. If
- * any of them use Data Transfer Objects (DTO), you can add them here as well.
- * </p>
- */
-@NullMarked
-//#if ui.framework == "hilla"
-@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
-//#endif
-package com.example.application.taskmanagement.service;
-
-import org.jspecify.annotations.NullMarked;
-//#if ui.framework == "hilla"
-import org.springframework.lang.NonNullApi;
-//#endif
